fix(preloader): fixes the 404 not found issue
this commit replaces the referenced image with a base64 inline version.
the reason is our current webpack setup parses the less and injects a
stripped down version into index.html, this stripped down version isnt
compatible with webpoack, because it references the original image
path - not the one creates by webpack.